fix [BUG]: REBUILD DB  #5
